Ensure that only your playlists are affected.
Because I want to be able to do this for playlists that aren't owned by
the caller, I have to use the Search API, and because the Search API
isn't guaranteed to return the caller's playlists before others, it's
possible for a completely random playlist to be returned while searching
for the `managed_pl`.

This commit adds a new `.env` file option, `SURFACE_OWNED_PLAYLIST`,
that will force the program to use your playlists. The implementation
could be better, but it works for now.